--- a/src/core/Akka.MultiNodeTestRunner/Program.cs
+++ b/src/core/Akka.MultiNodeTestRunner/Program.cs
@@ -343,7 +343,7 @@ namespace Akka.MultiNodeTestRunner
 
             Receive<Tcp.Received>(received =>
             {
-                var message = received.Data.DecodeString();
+                var message = received.Data.ToString();
                 _sinkCoordinator.Tell(message);
             });
         }