--- a/src/core/Akka.Streams.Tests/Dsl/JsonFramingSpec.cs
+++ b/src/core/Akka.Streams.Tests/Dsl/JsonFramingSpec.cs
@@ -46,7 +46,7 @@ namespace Akka.Streams.Tests.Dsl
                 .Via(JsonFraming.ObjectScanner(int.MaxValue))
                 .RunAggregate(new List<string>(), (list, s) =>
                 {
-                    list.Add(s.DecodeString());
+                    list.Add(s.ToString());
                     return list;
                 }, Materializer);
 
@@ -71,7 +71,7 @@ namespace Akka.Streams.Tests.Dsl
                 .Take(1)
                 .RunAggregate(new List<string>(), (list, s) =>
                 {
-                    list.Add(s.DecodeString());
+                    list.Add(s.ToString());
                     return list;
                 }, Materializer);
 
@@ -91,7 +91,7 @@ namespace Akka.Streams.Tests.Dsl
                 .Via(JsonFraming.ObjectScanner(int.MaxValue))
                 .RunAggregate(new List<string>(), (list, s) =>
                 {
-                    list.Add(s.DecodeString());
+                    list.Add(s.ToString());
                     return list;
                 }, Materializer);
 
@@ -114,7 +114,7 @@ namespace Akka.Streams.Tests.Dsl
                 .Via(JsonFraming.ObjectScanner(int.MaxValue))
                 .RunAggregate(new List<string>(), (list, s) =>
                 {
-                    list.Add(s.DecodeString());
+                    list.Add(s.ToString());
                     return list;
                 }, Materializer);
 
@@ -141,12 +141,13 @@ namespace Akka.Streams.Tests.Dsl
                 .Via(JsonFraming.ObjectScanner(int.MaxValue))
                 .RunAggregate(new List<string>(), (list, s) =>
                 {
-                    list.Add(s.DecodeString());
+                    list.Add(s.ToString());
                     return list;
-                }, Materializer);
+                }, Materializer)
+                .AwaitResult();
 
 
-            result.AwaitResult().ShouldAllBeEquivalentTo(new[]
+            result.ShouldAllBeEquivalentTo(new[]
             {
                 @"{ ""name"" : ""john"" }",
                 @"{ ""name"" : ""jack"" }"
@@ -162,7 +163,7 @@ namespace Akka.Streams.Tests.Dsl
             var result =
                 Source.FromPublisher(input)
                     .Via(JsonFraming.ObjectScanner(int.MaxValue))
-                    .Select(b => b.DecodeString())
+                    .Select(b => b.ToString())
                     .RunWith(Sink.FromSubscriber(output), Materializer);
 
             output.Request(1);
@@ -192,7 +193,7 @@ namespace Akka.Streams.Tests.Dsl
         {
             var buffer = new JsonObjectParser();
             buffer.Offer(ByteString.FromString("{}"));
-            buffer.Poll().Value.DecodeString().Should().Be("{}");
+            buffer.Poll().Value.ToString().Should().Be("{}");
         }
 
         [Fact]
@@ -200,7 +201,7 @@ namespace Akka.Streams.Tests.Dsl
         {
             var buffer = new JsonObjectParser();
             buffer.Offer(ByteString.FromString("{ \"name\": \"john\" }"));
-            buffer.Poll().Value.DecodeString().Should().Be("{ \"name\": \"john\" }");
+            buffer.Poll().Value.ToString().Should().Be("{ \"name\": \"john\" }");
         }
 
         [Fact] 
@@ -208,7 +209,7 @@ namespace Akka.Streams.Tests.Dsl
         {
             var buffer = new JsonObjectParser();
             buffer.Offer(ByteString.FromString("{ \"name\": \"john doe\" }"));
-            buffer.Poll().Value.DecodeString().Should().Be("{ \"name\": \"john doe\" }");
+            buffer.Poll().Value.ToString().Should().Be("{ \"name\": \"john doe\" }");
         }
 
         [Fact]
@@ -216,7 +217,7 @@ namespace Akka.Streams.Tests.Dsl
         {
             var buffer = new JsonObjectParser();
             buffer.Offer(ByteString.FromString("{ \"name\": \"john o'doe\" }"));
-            buffer.Poll().Value.DecodeString().Should().Be("{ \"name\": \"john o'doe\" }");
+            buffer.Poll().Value.ToString().Should().Be("{ \"name\": \"john o'doe\" }");
         }
 
         [Fact]
@@ -228,7 +229,7 @@ namespace Akka.Streams.Tests.Dsl
             buffer.Offer(ByteString.FromString("}"));
             buffer.Offer(ByteString.FromString("\""));
             buffer.Offer(ByteString.FromString("}"));
-            buffer.Poll().Value.DecodeString().Should().Be("{ \"name\": \"john{}\"}");
+            buffer.Poll().Value.ToString().Should().Be("{ \"name\": \"john{}\"}");
         }
 
         [Fact]
@@ -244,7 +245,7 @@ namespace Akka.Streams.Tests.Dsl
             buffer.Offer(ByteString.FromString("hey"));
             buffer.Offer(ByteString.FromString("\""));
             buffer.Offer(ByteString.FromString("}"));
-            buffer.Poll().Value.DecodeString().Should().Be("{ \"name\": \"john\\\"{}\\\" hey\"}");
+            buffer.Poll().Value.ToString().Should().Be("{ \"name\": \"john\\\"{}\\\" hey\"}");
         }
 
         [Fact]
@@ -252,7 +253,7 @@ namespace Akka.Streams.Tests.Dsl
         {
             var buffer = new JsonObjectParser();
             buffer.Offer(ByteString.FromString("{ \"age\" : 101}"));
-            buffer.Poll().Value.DecodeString().Should().Be("{ \"age\" : 101}");
+            buffer.Poll().Value.ToString().Should().Be("{ \"age\" : 101}");
         }
 
         [Fact]
@@ -260,7 +261,7 @@ namespace Akka.Streams.Tests.Dsl
         {
             var buffer = new JsonObjectParser();
             buffer.Offer(ByteString.FromString("{ \"age\" : 10.1}"));
-            buffer.Poll().Value.DecodeString().Should().Be("{ \"age\" : 10.1}");
+            buffer.Poll().Value.ToString().Should().Be("{ \"age\" : 10.1}");
         }
 
         [Fact]
@@ -275,7 +276,7 @@ namespace Akka.Streams.Tests.Dsl
                                    "  }" +
                                    "}";
             buffer.Offer(ByteString.FromString(content));
-            buffer.Poll().Value.DecodeString().Should().Be(content);
+            buffer.Poll().Value.ToString().Should().Be(content);
         }
 
         [Fact]
@@ -293,7 +294,7 @@ namespace Akka.Streams.Tests.Dsl
                                    "  }" +
                                    "}";
             buffer.Offer(ByteString.FromString(content));
-            buffer.Poll().Value.DecodeString().Should().Be(content);
+            buffer.Poll().Value.ToString().Should().Be(content);
         }
 
         [Fact]
@@ -330,7 +331,7 @@ namespace Akka.Streams.Tests.Dsl
                                    "  ]" +
                                    "}";
             buffer.Offer(ByteString.FromString(content));
-            buffer.Poll().Value.DecodeString().Should().Be(content);
+            buffer.Poll().Value.ToString().Should().Be(content);
         }
 
         [Fact]
@@ -360,7 +361,7 @@ namespace Akka.Streams.Tests.Dsl
                  ]
                }";
             buffer.Offer(ByteString.FromString(content));
-            buffer.Poll().Value.DecodeString().Should().Be(content);
+            buffer.Poll().Value.ToString().Should().Be(content);
         }
 
         [Fact]
@@ -369,7 +370,7 @@ namespace Akka.Streams.Tests.Dsl
             var buffer = new JsonObjectParser();
             const string content = "{ \"name\": \"john\", \"age\" : 101}";
             buffer.Offer(ByteString.FromString(content));
-            buffer.Poll().Value.DecodeString().Should().Be(content);
+            buffer.Poll().Value.ToString().Should().Be(content);
         }
 
         [Fact]
@@ -381,7 +382,7 @@ namespace Akka.Streams.Tests.Dsl
                                    "{ \"name\": \"john\"" +
                                    ", \"age\" : 101}";
             buffer.Offer(ByteString.FromString(content));
-            buffer.Poll().Value.DecodeString().Should().Be(content.TrimStart());
+            buffer.Poll().Value.ToString().Should().Be(content.TrimStart());
         }
 
         [Fact]
@@ -397,15 +398,15 @@ namespace Akka.Streams.Tests.Dsl
                  ""age"": 25
                }";
             buffer.Offer(ByteString.FromString(input1 + "," + input2));
-            buffer.Poll().Value.DecodeString().Should().Be(input1);
-            buffer.Poll().Value.DecodeString().Should().Be(input2);
+            buffer.Poll().Value.ToString().Should().Be(input1);
+            buffer.Poll().Value.ToString().Should().Be(input2);
 
             buffer.Poll().Should().Be(Option<ByteString>.None);
             buffer.Offer(ByteString.FromString("{\"name\":\"jenkins\",\"age\": "));
             buffer.Poll().Should().Be(Option<ByteString>.None);
 
             buffer.Offer(ByteString.FromString("65 }"));
-            buffer.Poll().Value.DecodeString().Should().Be("{\"name\":\"jenkins\",\"age\": 65 }");
+            buffer.Poll().Value.ToString().Should().Be("{\"name\":\"jenkins\",\"age\": 65 }");
         }
 
         [Fact]
@@ -446,7 +447,7 @@ namespace Akka.Streams.Tests.Dsl
 
             var result = Source.Single(ByteString.FromString(input))
                 .Via(JsonFraming.ObjectScanner(5))
-                .Select(b => b.DecodeString())
+                .Select(b => b.ToString())
                 .RunAggregate(new List<string>(), (list, s) =>
                 {
                     list.Add(s);