--- a/src/core/Akka.Streams/Dsl/Tcp.cs
+++ b/src/core/Akka.Streams/Dsl/Tcp.cs
@@ -294,4 +294,3 @@ namespace Akka.Streams.Dsl
         public TimeSpan Duration { get; }
     }
 }
-#endif