Continued work started at #2405 - now targeting v1.3 / .NET Core branch. Remaining bugs/features to take a look on - probably they will be implemented in upcoming PRs, they shouldn't delay this one from merging:

- [x] Port https://github.com/akka/akka/pull/21685
- [ ] `WriteFile` support - at the present moment .NET Core doesn't support TransferFile API at all. We could implement that as copying filestream to socket, however this is not what transfer file is all about.
- [x] Harden behavior under high amount of concurrent actors: `HorizontalScaleSpec` shown that at some point (around 550-600 concurrent clients on my laptop) TcpManager is failing. I don't know the exact reason yet, but it's deterministic, maybe related to some limits in configuration. Also this amount is already better that the current one (which caused my machine to boil on that amount of actors).
- [x] Investigate `Socket.SendAsync` method with BufferList - as `ByteString` implements byte buffers as array of `ArraySegment<byte>`, maybe we don't need to send those buffers one by one, maybe we could just supply buffer enumeration up front and leave rest to the .NET runtime.
- [ ] Investigate zero-copy buffers - while writing is copy free from our side, reading from socket is copying operation: we copy `SocketAsyncEventArgs` buffer to new `ByteString` instance, as we cannot guarantee that user will free that resource and when he will do that. Maybe there is some way to avoid that.
- [ ] Investigate [Windows RIO](https://technet.microsoft.com/en-us/library/hh997032(v=ws.11).aspx) - we could provide a windows specific implementation of TCP and UDP connection actors that would take advantage of registered I/O.