@dhwed 2017-05-19
I'm currently evaluating Akka.NET, and the TCP support in 1.2 has an issue.  If I abort the connection, wait for it to be re-established, and then abort it again, my TCP listener actor does not receive any subsequent Tcp.Connected messages (I've confirmed with Wireshark that the TCP connection is being re-established after the second RST).

After some minor refactoring to work against this pull request instead, I haven't seen any issues.  Hopefully this gets merged in soon.

Also, my code is much cleaner now that I can use the `ByteString.Buffers` property to get an `IEnumerable<ArraySegment<byte>>`.