@Horusiath 2017-05-23
We cannot do anything with mono test failures: the reason for bug lies inside Mono implementation of `SocketAsyncEventArgs` and `Socket` itself. Those tests should be muted. I can also see some racy test on Windows but it's not related to this PR.