@alexvaluyskiy 2017-05-23
@Horusiath could you mute them?