@alexvaluyskiy 2017-05-28
@Horusiath look at the artifact for `netcore` http://petabridge-ci.cloudapp.net/repository/download/AkkaNet_AkkaNetLinuxMonoUnitTests/27401:id/TestResults/Akka.Tests_xunit-netcoreapp1.1.xml