@Aaronontheweb 2017-06-23
Resolved differences with the upstream `v1.3` stuff