@Aaronontheweb 2017-06-24
my bad @Horusiath - used the web UI for doing the conflict resolution and screwed up. Should be fixed now.