@Aaronontheweb 2017-06-24
Sorry for the commit spam - even after the merge conflict was resolved there were a few things that still needed cleaning up. Think I've addressed them all now.