@Horusiath 2017-07-14
Build is failing (failed to clean tests) X-th time in a row. Anyone knows what's the reason?