@Horusiath 2017-07-23
### Some benchmarks

I've tested a simple client-server ping/pong scenarios with 50B messages. A single server and 2000 active concurrent clients talking on localhost (so yes, those numbers are totally artificial). OS: Windows 10 (x64).

| Server | Avg. msg/sec | Baseline |
|-----|-----|----|
| DotNetty | 40215,1 | 1.0 |
| Akka.Streams.IO (.NET) |  36983,65 | 0.92 |
| Akka.Streams.IO (JVM) | 37560 | 0.93 |